fix build script

+++ build.sh
@@ -8,7 +8,10 @@ cd ${DIR}
 
 rm -rf _book
 
-npm install markdown-parser2
+if [ ! -e node_modules ]; then
+  npm install markdown-parser2
+  gitbook install --log error
+fi
 
 for d in */; do
 
@@ -19,13 +22,16 @@ for d in */; do
   echo "Generate PDF for book ${d}"
   cp ./gen_summary.js ${d}
   cd ${d} && node gen_summary.js && rm gen_summary.js && cd ..
-  test ! -e node_modules && gitbook install --log error ${d}
+
+  if [ ! -e ${d}/node_modules ]; then
+    gitbook install --log error ${d}
+  fi
+
   gitbook pdf --log warn ${d} "${d}$(basename ${d}).pdf"
 done
 
 echo "Generate website"
 
-test ! -e node_modules && gitbook install --log error
 node gen_summary.js
 gitbook build --log warn
